solidity-verifiers: move tests to their corresponding files
