+++ solidity-verifiers/src/verifiers/g16.rs
@@ -70,3 +70,75 @@ impl ProtocolData for Groth16Data {
             .into_bytes()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        evm::{compile_solidity, save_solidity, Evm},
+        Groth16Data, ProtocolData,
+    };
+    use ark_bn254::{Bn254, Fr};
+    use ark_crypto_primitives::snark::SNARK;
+    use ark_ec::AffineRepr;
+    use ark_ff::{BigInt, BigInteger, PrimeField};
+    use ark_groth16::Groth16;
+    use ark_std::rand::{RngCore, SeedableRng};
+    use ark_std::test_rng;
+    use askama::Template;
+    use itertools::chain;
+
+    use super::Groth16Verifier;
+    use crate::verifiers::tests::{setup, DEFAULT_SETUP_LEN};
+
+    pub const FUNCTION_SELECTOR_GROTH16_VERIFY_PROOF: [u8; 4] = [0x43, 0x75, 0x3b, 0x4d];
+
+    #[test]
+    fn groth16_data_serde_roundtrip() {
+        let (_, _, _, vk, _) = setup(DEFAULT_SETUP_LEN);
+
+        let g16_data = Groth16Data::from(vk);
+        let mut bytes = vec![];
+        g16_data.serialize_protocol_data(&mut bytes).unwrap();
+        let obtained_g16_data = Groth16Data::deserialize_protocol_data(bytes.as_slice()).unwrap();
+
+        assert_eq!(g16_data, obtained_g16_data)
+    }
+
+    #[test]
+    fn test_groth16_verifier_accepts_and_rejects_proofs() {
+        let mut rng = ark_std::rand::rngs::StdRng::seed_from_u64(test_rng().next_u64());
+        let (_, _, g16_pk, g16_vk, circuit) = setup(DEFAULT_SETUP_LEN);
+        let g16_data = Groth16Data::from(g16_vk);
+
+        let proof = Groth16::<Bn254>::prove(&g16_pk, circuit, &mut rng).unwrap();
+        let res = Groth16Verifier::from(g16_data).render().unwrap();
+        save_solidity("groth16_verifier.sol", &res);
+        let groth16_verifier_bytecode = compile_solidity(&res, "Verifier");
+        let mut evm = Evm::default();
+        let verifier_address = evm.create(groth16_verifier_bytecode);
+        let (a_x, a_y) = proof.a.xy().unwrap();
+        let (b_x, b_y) = proof.b.xy().unwrap();
+        let (c_x, c_y) = proof.c.xy().unwrap();
+        let mut calldata: Vec<u8> = chain![
+            FUNCTION_SELECTOR_GROTH16_VERIFY_PROOF,
+            a_x.into_bigint().to_bytes_be(),
+            a_y.into_bigint().to_bytes_be(),
+            b_x.c1.into_bigint().to_bytes_be(),
+            b_x.c0.into_bigint().to_bytes_be(),
+            b_y.c1.into_bigint().to_bytes_be(),
+            b_y.c0.into_bigint().to_bytes_be(),
+            c_x.into_bigint().to_bytes_be(),
+            c_y.into_bigint().to_bytes_be(),
+            BigInt::from(Fr::from(circuit.z)).to_bytes_be(),
+        ]
+        .collect();
+        let (_, output) = evm.call(verifier_address, calldata.clone());
+        assert_eq!(*output.last().unwrap(), 1);
+
+        // change calldata to make it invalid
+        let last_calldata_element = calldata.last_mut().unwrap();
+        *last_calldata_element = 0;
+        let (_, output) = evm.call(verifier_address, calldata);
+        assert_eq!(*output.last().unwrap(), 0);
+    }
+}
