+++ solidity-verifiers/src/verifiers/kzg.rs
@@ -70,3 +70,132 @@ impl ProtocolData for KzgData {
             .into_bytes()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        evm::{compile_solidity, Evm},
+        utils::HeaderInclusion,
+        KzgData, ProtocolData,
+    };
+    use ark_bn254::{Bn254, Fr, G1Projective as G1};
+    use ark_ec::{AffineRepr, CurveGroup};
+    use ark_ff::{BigInteger, PrimeField};
+    use ark_std::rand::{RngCore, SeedableRng};
+    use ark_std::Zero;
+    use ark_std::{test_rng, UniformRand};
+    use askama::Template;
+    use itertools::chain;
+
+    use folding_schemes::{
+        commitment::{kzg::KZG, CommitmentScheme},
+        transcript::{
+            poseidon::{poseidon_test_config, PoseidonTranscript},
+            Transcript,
+        },
+    };
+
+    use super::KZG10Verifier;
+    use crate::verifiers::tests::{setup, DEFAULT_SETUP_LEN};
+
+    const FUNCTION_SELECTOR_KZG10_CHECK: [u8; 4] = [0x9e, 0x78, 0xcc, 0xf7];
+
+    #[test]
+    fn kzg_data_serde_roundtrip() {
+        let (pk, vk, _, _, _) = setup(DEFAULT_SETUP_LEN);
+
+        let kzg_data = KzgData::from((vk, Some(pk.powers_of_g[0..3].to_vec())));
+        let mut bytes = vec![];
+        kzg_data.serialize_protocol_data(&mut bytes).unwrap();
+        let obtained_kzg_data = KzgData::deserialize_protocol_data(bytes.as_slice()).unwrap();
+
+        assert_eq!(kzg_data, obtained_kzg_data)
+    }
+
+    #[test]
+    fn kzg_verifier_template_renders() {
+        let (kzg_pk, kzg_vk, _, _, _) = setup(DEFAULT_SETUP_LEN);
+        let kzg_data = KzgData::from((kzg_vk.clone(), Some(kzg_pk.powers_of_g[0..3].to_vec())));
+
+        let res = HeaderInclusion::<KZG10Verifier>::builder()
+            .template(kzg_data)
+            .build()
+            .render()
+            .unwrap();
+
+        // TODO: Unsure what this is testing. If we want to test correct rendering,
+        // we should first check that it COMPLETELY renders to what we expect.
+        assert!(res.contains(&kzg_vk.g.x.to_string()));
+    }
+
+    #[test]
+    fn kzg_verifier_compiles() {
+        let (kzg_pk, kzg_vk, _, _, _) = setup(DEFAULT_SETUP_LEN);
+        let kzg_data = KzgData::from((kzg_vk.clone(), Some(kzg_pk.powers_of_g[0..3].to_vec())));
+
+        let res = HeaderInclusion::<KZG10Verifier>::builder()
+            .template(kzg_data)
+            .build()
+            .render()
+            .unwrap();
+
+        let kzg_verifier_bytecode = compile_solidity(res, "KZG10");
+        let mut evm = Evm::default();
+        _ = evm.create(kzg_verifier_bytecode);
+    }
+
+    #[test]
+    fn kzg_verifier_accepts_and_rejects_proofs() {
+        let mut rng = ark_std::rand::rngs::StdRng::seed_from_u64(test_rng().next_u64());
+        let poseidon_config = poseidon_test_config::<Fr>();
+        let transcript_p = &mut PoseidonTranscript::<G1>::new(&poseidon_config);
+        let transcript_v = &mut PoseidonTranscript::<G1>::new(&poseidon_config);
+
+        let (kzg_pk, kzg_vk, _, _, _) = setup(DEFAULT_SETUP_LEN);
+        let kzg_data = KzgData::from((kzg_vk.clone(), Some(kzg_pk.powers_of_g[0..3].to_vec())));
+
+        let v: Vec<Fr> = std::iter::repeat_with(|| Fr::rand(&mut rng))
+            .take(DEFAULT_SETUP_LEN)
+            .collect();
+        let cm = KZG::<Bn254>::commit(&kzg_pk, &v, &Fr::zero()).unwrap();
+        let proof = KZG::<Bn254>::prove(&kzg_pk, transcript_p, &cm, &v, &Fr::zero(), None).unwrap();
+        let template = HeaderInclusion::<KZG10Verifier>::builder()
+            .template(kzg_data)
+            .build()
+            .render()
+            .unwrap();
+
+        let kzg_verifier_bytecode = compile_solidity(template, "KZG10");
+        let mut evm = Evm::default();
+        let verifier_address = evm.create(kzg_verifier_bytecode);
+
+        let (cm_affine, proof_affine) = (cm.into_affine(), proof.proof.into_affine());
+        let (x_comm, y_comm) = cm_affine.xy().unwrap();
+        let (x_proof, y_proof) = proof_affine.xy().unwrap();
+        let y = proof.eval.into_bigint().to_bytes_be();
+
+        transcript_v.absorb_point(&cm).unwrap();
+        let x = transcript_v.get_challenge();
+
+        let x = x.into_bigint().to_bytes_be();
+        let mut calldata: Vec<u8> = chain![
+            FUNCTION_SELECTOR_KZG10_CHECK,
+            x_comm.into_bigint().to_bytes_be(),
+            y_comm.into_bigint().to_bytes_be(),
+            x_proof.into_bigint().to_bytes_be(),
+            y_proof.into_bigint().to_bytes_be(),
+            x.clone(),
+            y,
+        ]
+        .collect();
+
+        let (_, output) = evm.call(verifier_address, calldata.clone());
+        assert_eq!(*output.last().unwrap(), 1);
+
+        // change calldata to make it invalid
+        let last_calldata_element = calldata.last_mut().unwrap();
+        *last_calldata_element = 0;
+        let (_, output) = evm.call(verifier_address, calldata);
+        assert_eq!(*output.last().unwrap(), 0);
+    }
+}
